// src/stats/types/team-dashboard-by-shooting-splits.ts
export interface TeamDashboardByShootingSplitsParams {
  teamID: number;
  lastNGames?: number;
  measureType?: string;
  month?: number;
  opponentTeamID?: number;
  paceAdjust?: string;
  perMode?: string;
  period?: number;
  plusMinus?: string;
  rank?: string;
  season?: string;
  seasonType?: string;
  dateFrom?: string;
  dateTo?: string;
  gameSegment?: string;
  leagueID?: string;
  location?: string;
  outcome?: string;
  pORound?: number;
  seasonSegment?: string;
  shotClockRange?: number;
  vsConference?: string;
  vsDivision?: string;
}

export interface AssistedByRow {
  groupSet: unknown;
  playerId: number;
  playerName: string;
  fgm: number;
  fga: number;
  fgPct: number | null;
  fg3m: number;
  fg3a: number;
  fg3Pct: number | null;
  efgPct: number | null;
  blka: number;
  pctAst2pm: unknown;
  pctUast2pm: unknown;
  pctAst3pm: unknown;
  pctUast3pm: unknown;
  pctAstFgm: unknown;
  pctUastFgm: unknown;
  fgmRank: number;
  fgaRank: number;
  fgPctRank: number;
  fg3mRank: number;
  fg3aRank: number;
  fg3PctRank: number;
  efgPctRank: number;
  blkaRank: number;
  pctAst2pmRank: number;
  pctUast2pmRank: number;
  pctAst3pmRank: number;
  pctUast3pmRank: number;
  pctAstFgmRank: number;
  pctUastFgmRank: number;
  cfid: number;
  cfparams: unknown;
}

export interface AssitedShotTeamDashboardRow {
  groupSet: unknown;
  groupValue: unknown;
  fgm: number;
  fga: number;
  fgPct: number | null;
  fg3m: number;
  fg3a: number;
  fg3Pct: number | null;
  efgPct: number | null;
  blka: number;
  pctAst2pm: unknown;
  pctUast2pm: unknown;
  pctAst3pm: unknown;
  pctUast3pm: unknown;
  pctAstFgm: unknown;
  pctUastFgm: unknown;
  fgmRank: number;
  fgaRank: number;
  fgPctRank: number;
  fg3mRank: number;
  fg3aRank: number;
  fg3PctRank: number;
  efgPctRank: number;
  blkaRank: number;
  pctAst2pmRank: number;
  pctUast2pmRank: number;
  pctAst3pmRank: number;
  pctUast3pmRank: number;
  pctAstFgmRank: number;
  pctUastFgmRank: number;
  cfid: number;
  cfparams: unknown;
}

export interface OverallTeamDashboardRow {
  groupSet: unknown;
  groupValue: unknown;
  fgm: number;
  fga: number;
  fgPct: number | null;
  fg3m: number;
  fg3a: number;
  fg3Pct: number | null;
  efgPct: number | null;
  blka: number;
  pctAst2pm: unknown;
  pctUast2pm: unknown;
  pctAst3pm: unknown;
  pctUast3pm: unknown;
  pctAstFgm: unknown;
  pctUastFgm: unknown;
  fgmRank: number;
  fgaRank: number;
  fgPctRank: number;
  fg3mRank: number;
  fg3aRank: number;
  fg3PctRank: number;
  efgPctRank: number;
  blkaRank: number;
  pctAst2pmRank: number;
  pctUast2pmRank: number;
  pctAst3pmRank: number;
  pctUast3pmRank: number;
  pctAstFgmRank: number;
  pctUastFgmRank: number;
  cfid: number;
  cfparams: unknown;
}

export interface Shot5FTTeamDashboardRow {
  groupSet: unknown;
  groupValue: unknown;
  fgm: number;
  fga: number;
  fgPct: number | null;
  fg3m: number;
  fg3a: number;
  fg3Pct: number | null;
  efgPct: number | null;
  blka: number;
  pctAst2pm: unknown;
  pctUast2pm: unknown;
  pctAst3pm: unknown;
  pctUast3pm: unknown;
  pctAstFgm: unknown;
  pctUastFgm: unknown;
  fgmRank: number;
  fgaRank: number;
  fgPctRank: number;
  fg3mRank: number;
  fg3aRank: number;
  fg3PctRank: number;
  efgPctRank: number;
  blkaRank: number;
  pctAst2pmRank: number;
  pctUast2pmRank: number;
  pctAst3pmRank: number;
  pctUast3pmRank: number;
  pctAstFgmRank: number;
  pctUastFgmRank: number;
  cfid: number;
  cfparams: unknown;
}

export interface Shot8FTTeamDashboardRow {
  groupSet: unknown;
  groupValue: unknown;
  fgm: number;
  fga: number;
  fgPct: number | null;
  fg3m: number;
  fg3a: number;
  fg3Pct: number | null;
  efgPct: number | null;
  blka: number;
  pctAst2pm: unknown;
  pctUast2pm: unknown;
  pctAst3pm: unknown;
  pctUast3pm: unknown;
  pctAstFgm: unknown;
  pctUastFgm: unknown;
  fgmRank: number;
  fgaRank: number;
  fgPctRank: number;
  fg3mRank: number;
  fg3aRank: number;
  fg3PctRank: number;
  efgPctRank: number;
  blkaRank: number;
  pctAst2pmRank: number;
  pctUast2pmRank: number;
  pctAst3pmRank: number;
  pctUast3pmRank: number;
  pctAstFgmRank: number;
  pctUastFgmRank: number;
  cfid: number;
  cfparams: unknown;
}

export interface ShotAreaTeamDashboardRow {
  groupSet: unknown;
  groupValue: unknown;
  fgm: number;
  fga: number;
  fgPct: number | null;
  fg3m: number;
  fg3a: number;
  fg3Pct: number | null;
  efgPct: number | null;
  blka: number;
  pctAst2pm: unknown;
  pctUast2pm: unknown;
  pctAst3pm: unknown;
  pctUast3pm: unknown;
  pctAstFgm: unknown;
  pctUastFgm: unknown;
  fgmRank: number;
  fgaRank: number;
  fgPctRank: number;
  fg3mRank: number;
  fg3aRank: number;
  fg3PctRank: number;
  efgPctRank: number;
  blkaRank: number;
  pctAst2pmRank: number;
  pctUast2pmRank: number;
  pctAst3pmRank: number;
  pctUast3pmRank: number;
  pctAstFgmRank: number;
  pctUastFgmRank: number;
  cfid: number;
  cfparams: unknown;
}

export interface ShotTypeTeamDashboardRow {
  groupSet: unknown;
  groupValue: unknown;
  fgm: number;
  fga: number;
  fgPct: number | null;
  fg3m: number;
  fg3a: number;
  fg3Pct: number | null;
  efgPct: number | null;
  blka: number;
  pctAst2pm: unknown;
  pctUast2pm: unknown;
  pctAst3pm: unknown;
  pctUast3pm: unknown;
  pctAstFgm: unknown;
  pctUastFgm: unknown;
  fgmRank: number;
  fgaRank: number;
  fgPctRank: number;
  fg3mRank: number;
  fg3aRank: number;
  fg3PctRank: number;
  efgPctRank: number;
  blkaRank: number;
  pctAst2pmRank: number;
  pctUast2pmRank: number;
  pctAst3pmRank: number;
  pctUast3pmRank: number;
  pctAstFgmRank: number;
  pctUastFgmRank: number;
  cfid: number;
  cfparams: unknown;
}

export interface TeamDashboardByShootingSplitsResponse {
  assistedBy: AssistedByRow[];
  assitedShotTeamDashboard: AssitedShotTeamDashboardRow[];
  overallTeamDashboard: OverallTeamDashboardRow[];
  shot5FTTeamDashboard: Shot5FTTeamDashboardRow[];
  shot8FTTeamDashboard: Shot8FTTeamDashboardRow[];
  shotAreaTeamDashboard: ShotAreaTeamDashboardRow[];
  shotTypeTeamDashboard: ShotTypeTeamDashboardRow[];
}